[C++ification] Take care of a few warnings and don't use LTO
Desktop compilers mis-detect memory allocation size in our bundled property
code if fortified version of `strncpy` is used (even though the size
calculation is correct). Use `memcpy` instead.

LTO doesn't appear to be ready for use just yet. It gives warnings similar to
these:

    jni-wrappers.hh(143,41): warning GD616DC71: argument 1 value ‘18446744073709551615’ exceeds maximum object size 9223372036854775807
    util.cc(75,8): warning GA5933D32: ‘MEM[(struct timing_point *)&partial_time].ns’ may be used uninitialized in this function [-Wmaybe-uninitialized]

Which, even though well-intentioned (and, technically, correct) are irrelevant
in our case and the compiler+linker diagnostics fail to understand the intention
and usage of the code.

Additionally, fix a few instances where type conversion is different for
different platforms/targets.

+++ src/monodroid/jni/monodroid-glue.cc
@@ -916,7 +916,7 @@ MonodroidRuntime::init_android_runtime (MonoDomain *domain, JNIEnv *env, jclass 
 	init.localRefsAreIndirect   = LocalRefsAreIndirect (env, runtimeClass, init.androidSdkVersion);
 	init.isRunningOnDesktop     = is_running_on_desktop ? 1 : 0;
 	init.brokenExceptionTransitions = application_config.broken_exception_transitions ? 1 : 0;
-	init.packageNamingPolicy    = application_config.package_naming_policy;
+	init.packageNamingPolicy    = static_cast<int>(application_config.package_naming_policy);
 
 	// GC threshold is 90% of the max GREF count
 	init.grefGcThreshold        = static_cast<int>(androidSystem.get_gref_gc_threshold ());
